Mergins Ismail's code and mine, improved paper selection logic.
