+++ pages/4_question_cascade.py
@@ -1,93 +1,219 @@
 import streamlit as st
+import os
+import json
+import pandas as pd
+from streamlit_cookies_manager import CookieManager
+from src.various import handle_redirects, get_selected_paper
 
-st.set_page_config(page_title="Questionnaire", layout="wide")
+# Set page config
+st.set_page_config(page_title="Browse Paper", layout="wide", initial_sidebar_state="collapsed")
 
-paper_title = "Authors2, Title2, Journal2, Issue2, volume2, pages2"
-full_text_url = "https://example.com/fulltext/2"
+# Initialize the cookie manager
+cookies = CookieManager(prefix="annotation_app_")
+if not cookies.ready():
+    st.stop()
 
-spacer_left, main_col, spacer_right = st.columns([1, 2, 1])
-with main_col:
-    st.markdown("### You have selected to annotate the paper:")
-    st.markdown(f"**{paper_title}**\n\n"
-                "Use the link below to scan through the full-text paper and then answer the quick questionnaire. "
-                "You can go back via “Pick another” as many times as you want. You’ll only be allowed to abandon a “Confirmed Paper” twice. "
-                "Once you’re happy, press “Confirm paper.”")
+handle_redirects(cookies)
 
-    st.markdown(
-        f"""<a href="{full_text_url}" target="_blank">
-                <button style="
-                  padding:8px 16px;
-                  background-color:#5A9;
-                  color:white;
-                  border:none;
-                  border-radius:4px;
-                  cursor:pointer;
-                ">
-                  Go to full-text paper
-                </button>
-            </a>""",
-        unsafe_allow_html=True,
+# Paths
+JSON_FOLDER = "Full_text_jsons"
+USERS_TABLE_PATH = "AWS_S3\\users_table.xlsx"
+
+# Function to load the selected paper's JSON file based on the PMID
+def load_paper_by_pmid(pmid):
+    for filename in os.listdir(JSON_FOLDER):
+        if filename.endswith(".json"):
+            try:
+                with open(os.path.join(JSON_FOLDER, filename), "r", encoding="utf-8") as f:
+                    raw = json.load(f)
+                    # Check if the PMID matches
+                    doc = raw[0]["documents"][0]
+                    front = doc["passages"][0]  # Front matter
+                    meta = front["infons"]
+                    extracted_pmid = meta.get("article-id_pmid", None)
+                    if extracted_pmid == pmid:
+                        return raw
+            except Exception as e:
+                st.error(f"Error reading file {filename}: {e}")
+    st.error(f"No JSON file found for PMID: {pmid}")
+    st.stop()
+
+# Function to update the "Paper in progress" column
+def update_paper_in_progress(user_id, pmid):
+    # Load the users table
+    users_df = pd.read_excel(USERS_TABLE_PATH)
+
+    # Find the row corresponding to the user
+    user_row = users_df[users_df["userID"] == user_id]
+
+    if not user_row.empty:
+        # Update the "Paper in progress" column
+        users_df.loc[users_df["userID"] == user_id, "Paper in progress"] = pmid
+
+        # Save the updated table back to the Excel file
+        users_df.to_excel(USERS_TABLE_PATH, index=False)
+
+        # Save the updated paper PMID in cookies and session state
+        st.session_state["paper_in_progress"] = pmid
+
+        cookies["paper_in_progress"] = pmid
+        cookies.save()
+    else:
+        print(f"User with ID {user_id} not found.")
+
+
+# Fetch the PMID
+pmid = get_selected_paper(cookies)
+if pmid is None:
+    st.error("No paper selected. Please select a paper to annotate.")
+    st.switch_page("pages/2_pick_paper.py")
+
+# Load the selected paper's JSON file
+raw = load_paper_by_pmid(pmid)
+doc = raw[0]["documents"][0]
+
+# Extract paper metadata
+front = doc["passages"][0]  # Front matter
+meta = front["infons"]
+title = front["text"]
+# Extract and clean up authors
+authors = []
+for k, v in meta.items():
+    if k.startswith("name_"):
+        parts = v.split(";")
+        surname = next((p.split(":")[1] for p in parts if p.startswith("surname:")), "").strip()
+        given_names = next((p.split(":")[1] for p in parts if p.startswith("given-names:")), "").strip()
+        if surname and given_names:
+            authors.append(f"{given_names} {surname}")
+        elif surname:
+            authors.append(surname)
+        elif given_names:
+            authors.append(given_names)
+authors_str = ", ".join(authors)
+
+# Dynamically construct the metadata parts
+metadata_parts = []
+if meta.get("issue", "?") != "?":
+    metadata_parts.append(f"**Issue:** {meta['issue']}")
+if meta.get("volume", "?") != "?":
+    metadata_parts.append(f"**Volume:** {meta['volume']}")
+fpage = meta.get("fpage", "N/A")
+lpage = meta.get("lpage", "N/A")
+if fpage != "N/A" and lpage != "N/A":
+    metadata_parts.append(f"**Pages:** {fpage}-{lpage}")
+year = meta.get("year", "?")
+if year != "?":
+    metadata_parts.append(f"**Year:** {year}")
+
+# Construct the DOI link if available
+doi = meta.get("article-id_doi", "")
+doi_link = f"https://doi.org/{doi}" if doi else None
+
+# Display paper metadata with reduced top margin
+st.markdown(f"""
+    <div style="margin-top: -50px;">
+        <h3>You have selected to annotate the paper:</h3>
+        <strong>{authors_str}</strong>, <em>{title}</em>
+    </div>
+""", unsafe_allow_html=True)
+if metadata_parts:
+    st.markdown(", ".join(metadata_parts))
+
+# Add a centered descriptive paragraph
+st.markdown("###")
+col1, col2, col3 = st.columns([0.5, 3, 0.5])  # Create three columns for centering
+with col2:  # Use the middle column
+    st.markdown("""
+        <div style="text-align: center; margin-bottom: -100px; margin-top: -40px;">
+            Use the link to the full-text paper to scan through it and then answer the quick questionnaire below. 
+            In this stage, you can go back through the "Pick another" button and change the paper as many times as you want. 
+            You will be allowed to abandon a "Confirmed Paper" later only twice. 
+            If you are happy with your selection, press "Confirm paper" to proceed to annotating your current selection.
+        </div>
+    """, unsafe_allow_html=True)
+
+# Center the "Go to full-text paper" button
+st.markdown("###")
+col1, col2, col3 = st.columns([1.5, 1, 1])  # Create three columns for centering
+with col2:  # Use the middle column
+    if doi_link:
+        st.button("Go to full-text paper", on_click=lambda: st.write(f"[Go to full-text paper]({doi_link})"))
+    else:
+        st.warning("DOI link not available for this paper.")
+# Questionnaire
+st.markdown("### Questionnaire")
+
+# Question 1
+st.markdown('<div>1. Is the paper describing wet lab experiments that aim to understand protein interactions?</div>', unsafe_allow_html=True)
+col1, col2 = st.columns([1, 200])  # Create two columns for better layout
+with col1:
+    st.markdown("")
+with col2:
+    q1 = st.radio("Question 1", options=["YES", "NO"], horizontal=True, key="q1_radio", label_visibility="collapsed")
+
+# Sub-question 1a
+st.markdown('<div style="padding-left: 20px; margin-bottom: 10px;">1a. What is the main method the authors use to understand protein interactions?</div>', unsafe_allow_html=True)
+col1, col2 = st.columns([1, 60])  # Create two columns for better layout
+with col1:
+    st.markdown("")
+with col2:
+    q1a = st.text_input("Sub-question 1a", placeholder="Enter the main method here", key="q1a_text", label_visibility="collapsed")
+
+# Sub-question 1b
+st.markdown('<div style="padding-left: 20px;">1b. Is this method preserving protein interactions in a cell-free system (e.g., whole cell extracts)?</div>', unsafe_allow_html=True)
+col1, col2 = st.columns([1, 50])  # Create two columns for better layout
+with col1:
+    st.markdown("")
+with col2:
+    q1b = st.radio("Sub-question 1b", options=["YES", "NO"], horizontal=True, key="q1b_radio", label_visibility="collapsed")
+
+# Sub-question 1c
+st.markdown('<div style="padding-left: 20px;">1c. Is this method using any type of cross-linking to preserve protein interactions?</div>', unsafe_allow_html=True)
+col1, col2 = st.columns([1, 50])  # Create two columns for better layout
+with col1:
+    st.markdown("")
+with col2:
+    q1c = st.radio("Sub-question 1c", options=["YES", "NO"], horizontal=True, key="q1c_radio", label_visibility="collapsed")
+
+# Question 2
+st.markdown('<div style="margin-bottom: 10px;">2. What is your level of familiarity with the topic of this paper?</div>', unsafe_allow_html=True)
+q2 = st.selectbox("Question 2", ["Basic", "Course", "MSc research", "PhD field", "PhD research", "Expert"], key="q2_select", label_visibility="collapsed")
+
+# Question 3
+st.markdown('<div style="margin-bottom: 10px;">3. What is your level of familiarity with the methods and experiments in this paper?</div>', unsafe_allow_html=True)
+q3 = st.selectbox("Question 3", ["Basic", "Course", "MSc research", "PhD field", "PhD research", "Expert"], key="q3_select", label_visibility="collapsed")
+
+# Navigation buttons
+st.markdown("""
+    <div style="margin-top: -20px;">
+    </div>
+""", unsafe_allow_html=True)
+
+# Check if all fields are filled
+all_filled = (
+    q1 is not None and
+    q1a.strip() != "" and
+    q1b is not None and
+    q1c is not None and
+    q2 is not None and
+    q3 is not None
+)
+
+col1, col2, col3 = st.columns([1.5, 1, 1])  # Adjust column widths to center the buttons properly
+with col1:  # Place "Pick another" button in the left column
+    if st.button("Pick another paper", type="secondary", key="pick_another_button"):
+        st.session_state["selected_paper"] = None
+        cookies["selected_paper"] = None
+        st.switch_page("pages/2_pick_paper.py")
+with col2:  # Use the middle column for "Confirm paper" button
+    st.button(
+        "Confirm paper",
+        type="primary",
+        key="confirm_paper_button",
+        disabled=not all_filled
     )
-
-    st.write("---")
-
-    # Q1: Yes/No
-    q1 = st.radio(
-        "1. Is the paper describing wet lab experiments that aim to understand protein interactions?",
-        ("Yes", "No"),
-        index=1,
-        key="q1"
-    )
-    disabled = (q1 == "No")
-
-    # 1a–c
-    st.markdown("**1a.** What is the main method the authors use to understand protein interactions?")
-    st.text_input(
-    label="Answer 1a (hidden)", 
-    placeholder="Type your answer here…",
-    disabled=disabled, 
-    key="q1a",
-    label_visibility="collapsed",)
-
-    st.markdown("**1b.** Is this method preserving protein interactions in a cell-free system (eg whole cell extracts)?")
-    st.radio(
-    label="Answer 1b (hidden)", 
-    options=("Yes","No"),
-    disabled=disabled,
-    key="q1b",
-    label_visibility="collapsed",)
-
-    st.markdown("**1c.** Is this method using any type of cross-linking to preserve protein interactions?")
-    st.radio(
-    label="Answer 1c (hidden)", 
-    options=("Yes","No"), 
-    disabled=disabled, 
-    key="q1c", 
-    label_visibility="collapsed")
-
-    # Q2 & Q3
-    levels = ["Basic", "MSc course", "MSc research", "PhD field", "PhD research", "Expert"]
-
-    st.selectbox(
-    "2. What is your level of familiarity with the topic of this paper?",
-    levels,
-    key="q2",
-    disabled=disabled)
-
-    st.selectbox(
-    "3. What is your level of familiarity with the methods and experiments in this paper?",
-    levels,
-    key="q3",
-    disabled=disabled)
-
-
-
-    st.write("")
-
-    sp1, btn_col1, gap, btn_col2, sp2 = st.columns([1, 2, 1.5, 2, 1])
-    with btn_col1:
-        if st.button("Pick another"):
-            st.switch_page("pages/2_pick_paper.py")
-    with btn_col2:
-        if st.button("Confirm paper", type="primary", disabled=disabled):
-            st.switch_page("pages/5_detail_picker.py")
+    if all_filled and st.session_state.get("confirm_paper_button"):
+        st.set_option("client.showSidebarNavigation", False)
+        # Update the "Paper in progress" column
+        update_paper_in_progress(st.session_state["userID"], pmid)
+        st.switch_page("pages/5_detail_picker.py")
